+++ js/level2.js
@@ -6,8 +6,8 @@ export function createLevel2(widthModifier,heightModifier,offscreen) {
   const level2 = createElement('div', 'level2', ['box'], {
     position: 'absolute',
     background: 'white',
-    width: '1000px',
-    height: '600px',
+    width: `${widthModifier * 500}px`,
+    height: `${heightModifier * 300}px`,
     padding: '0px',
     border: '0px',
     left: `${offscreen}%`,
@@ -162,355 +162,355 @@ export function createLevel2(widthModifier,heightModifier,offscreen) {
         top: `${heightModifier * 60}px`, 
         left: `${widthModifier * 320}px`
       },
-      { id: 'mazeWall23', 
+      { id: 'mazeWall24', 
         width: `${widthModifier * 130}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 100}px`, 
         left: `${widthModifier * 320}px`
       },
-      { id: 'mazeWall23', 
+      { id: 'mazeWall25', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 70}px`, 
         left: `${widthModifier * 450}px`
       },
-      { id: 'mazeWall23', 
+      { id: 'mazeWall26', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 70}px`, 
         left: `${widthModifier * 450}px`
       },
-      { id: 'mazeWall23', 
+      { id: 'mazeWall27', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 20}px`, 
         top: `${heightModifier * 70}px`, 
         left: `${widthModifier * 470}px`
       },
-      { id: 'mazeWall24', 
+      { id: 'mazeWall28', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 100}px`, 
         left: `${widthModifier * 480}px`
       },
-      { id: 'mazeWall25', 
+      { id: 'mazeWall29', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 120}px`, 
         left: `${widthModifier * 480}px`
       },
-      { id: 'mazeWall26', 
+      { id: 'mazeWall30', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 140}px`, 
         left: `${widthModifier * 480}px`
       },
-      { id: 'mazeWall27', 
+      { id: 'mazeWall31', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 160}px`, 
         left: `${widthModifier * 480}px`
       },
-      { id: 'mazeWall28', 
+      { id: 'mazeWall32', 
         width: `${widthModifier * 70}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 190}px`, 
         left: `${widthModifier * 420}px`
       },
-      { id: 'mazeWall29', 
+      { id: 'mazeWall33', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 160}px`, 
         left: `${widthModifier * 410}px`
       },
-      { id: 'mazeWall30', 
+      { id: 'mazeWall34', 
         width: `${widthModifier * 40}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 170}px`, 
         left: `${widthModifier * 430}px`
       },
-      { id: 'mazeWall31', 
+      { id: 'mazeWall35', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 120}px`, 
         left: `${widthModifier * 460}px`
       },
-      { id: 'mazeWall32', 
+      { id: 'mazeWall36', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 20}px`, 
         top: `${heightModifier * 140}px`, 
         left: `${widthModifier * 450}px`
       },
-      { id: 'mazeWall33', 
+      { id: 'mazeWall37', 
         width: `${widthModifier * 50}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 150}px`, 
         left: `${widthModifier * 400}px`
       },
-      { id: 'mazeWall34', 
+      { id: 'mazeWall38', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 120}px`, 
         left: `${widthModifier * 400}px`
       },
-      { id: 'mazeWall35', 
+      { id: 'mazeWall39', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 110}px`, 
         left: `${widthModifier * 430}px`
       },
-      { id: 'mazeWall35', 
+      { id: 'mazeWall40', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 110}px`, 
         left: `${widthModifier * 420}px`
       },
-      { id: 'mazeWall36', 
+      { id: 'mazeWall41', 
         width: `${widthModifier * 140}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 120}px`, 
         left: `${widthModifier * 260}px`
       },
-      { id: 'mazeWall37', 
+      { id: 'mazeWall42', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 120}px`, 
         top: `${heightModifier * 40}px`, 
         left: `${widthModifier * 250}px`
       },
-      { id: 'mazeWall38', 
+      { id: 'mazeWall43', 
         width: `${widthModifier * 70}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 80}px`, 
         left: `${widthModifier * 180}px`
       },
-      { id: 'mazeWall39', 
+      { id: 'mazeWall44', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 100}px`, 
         left: `${widthModifier * 230}px`
       },
-      { id: 'mazeWall40', 
+      { id: 'mazeWall45', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 140}px`, 
         left: `${widthModifier * 220}px`
       },
-      { id: 'mazeWall41', 
+      { id: 'mazeWall46', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 110}px`, 
         left: `${widthModifier * 210}px`
       },
-      { id: 'mazeWall42', 
+      { id: 'mazeWall47', 
         width: `${widthModifier * 60}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 100}px`, 
         left: `${widthModifier * 160}px`
       },
-      { id: 'mazeWall43', 
+      { id: 'mazeWall48', 
         width: `${widthModifier * 50}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 100}px`, 
         left: `${widthModifier * 100}px`
       },
-      { id: 'mazeWall44', 
+      { id: 'mazeWall49', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 50}px`, 
         top: `${heightModifier * 110}px`, 
         left: `${widthModifier * 140}px`
       },
-      { id: 'mazeWall45', 
+      { id: 'mazeWall50', 
         width: `${widthModifier * 40}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 120}px`, 
         left: `${widthModifier * 90}px`
       },
-      { id: 'mazeWall46', 
+      { id: 'mazeWall51', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 130}px`, 
         left: `${widthModifier * 120}px`
       },
-      { id: 'mazeWall47', 
+      { id: 'mazeWall52', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 60}px`, 
         top: `${heightModifier * 140}px`, 
         left: `${widthModifier * 100}px`
       },
-      { id: 'mazeWall48', 
+      { id: 'mazeWall53', 
         width: `${widthModifier * 40}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 170}px`, 
         left: `${widthModifier * 110}px`
       },
-      { id: 'mazeWall49', 
+      { id: 'mazeWall54', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 180}px`, 
         left: `${widthModifier * 140}px`
       },
-      { id: 'mazeWall50', 
+      { id: 'mazeWall55', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 180}px`, 
         left: `${widthModifier * 170}px`
       },
-      { id: 'mazeWall51', 
+      { id: 'mazeWall56', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 50}px`, 
         top: `${heightModifier * 180}px`, 
         left: `${widthModifier * 190}px`
       },
-      { id: 'mazeWall52', 
+      { id: 'mazeWall57', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 180}px`, 
         left: `${widthModifier * 200}px`
       },
-      { id: 'mazeWall53', 
+      { id: 'mazeWall58', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 20}px`, 
         top: `${heightModifier * 160}px`, 
         left: `${widthModifier * 210}px`
       },
-      { id: 'mazeWall54', 
+      { id: 'mazeWall59', 
         width: `${widthModifier * 40}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 160}px`, 
         left: `${widthModifier * 220}px`
       },
-      { id: 'mazeWall55', 
+      { id: 'mazeWall60', 
         width: `${widthModifier * 90}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 210}px`, 
         left: `${widthModifier * 90}px`
       },
-      { id: 'mazeWall56', 
+      { id: 'mazeWall61', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 20}px`, 
         top: `${heightModifier * 190}px`, 
         left: `${widthModifier * 120}px`
       },
-      { id: 'mazeWall57', 
+      { id: 'mazeWall62', 
         width: `${widthModifier * 50}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 230}px`, 
         left: `${widthModifier * 150}px`
       },
-      { id: 'mazeWall58', 
+      { id: 'mazeWall63', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 50}px`, 
         top: `${heightModifier * 230}px`, 
         left: `${widthModifier * 140}px`
       },
-      { id: 'mazeWall59', 
+      { id: 'mazeWall64', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 250}px`, 
         left: `${widthModifier * 160}px`
       },
-      { id: 'mazeWall60', 
+      { id: 'mazeWall65', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 240}px`, 
         left: `${widthModifier * 180}px`
       },
-      { id: 'mazeWall61', 
+      { id: 'mazeWall66', 
         width: `${widthModifier * 70}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 270}px`, 
         left: `${widthModifier * 180}px`
       },
-      { id: 'mazeWall62', 
+      { id: 'mazeWall67', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 240}px`, 
         left: `${widthModifier * 250}px`
       },
-      { id: 'mazeWall63', 
+      { id: 'mazeWall68', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 180}px`, 
         left: `${widthModifier * 250}px`
       },
-      { id: 'mazeWall63', 
+      { id: 'mazeWall69', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 180}px`, 
         left: `${widthModifier * 230}px`
       },
-      { id: 'mazeWall64', 
+      { id: 'mazeWall70', 
         width: `${widthModifier * 30}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 200}px`, 
         left: `${widthModifier * 210}px`
       },
-      { id: 'mazeWall65', 
+      { id: 'mazeWall71', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 50}px`, 
         top: `${heightModifier * 210}px`, 
         left: `${widthModifier * 230}px`
       },
-      { id: 'mazeWall66', 
+      { id: 'mazeWall72', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 30}px`, 
         top: `${heightModifier * 220}px`, 
         left: `${widthModifier * 210}px`
       },
-      { id: 'mazeWall67', 
+      { id: 'mazeWall73', 
         width: `${widthModifier * 20}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 250}px`, 
         left: `${widthModifier * 200}px`
       },
-      { id: 'mazeWall68', 
+      { id: 'mazeWall74', 
         width: `${widthModifier * 140}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 190}px`, 
         left: `${widthModifier * 260}px`
       },
-      { id: 'mazeWall69', 
+      { id: 'mazeWall75', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 90}px`, 
         top: `${heightModifier * 200}px`, 
         left: `${widthModifier * 350}px`
       },
-      { id: 'mazeWall70', 
+      { id: 'mazeWall76', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 20}px`, 
         top: `${heightModifier * 200}px`, 
         left: `${widthModifier * 430}px`
       },
-      { id: 'mazeWall71', 
+      { id: 'mazeWall77', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 40}px`, 
         top: `${heightModifier * 250}px`, 
         left: `${widthModifier * 430}px`
       },
-      { id: 'mazeWall72', 
+      { id: 'mazeWall78', 
         width: `${widthModifier * 30}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 250}px`, 
         left: `${widthModifier * 440}px`
       },
-      { id: 'mazeWall73', 
+      { id: 'mazeWall79', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 240}px`, 
         left: `${widthModifier * 460}px`
       },
-      { id: 'mazeWall74', 
+      { id: 'mazeWall80', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 200}px`, 
         left: `${widthModifier * 440}px`
       },
-      { id: 'mazeWall75', 
+      { id: 'mazeWall81', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 70}px`, 
         top: `${heightModifier * 210}px`, 
         left: `${widthModifier * 470}px`
       },
-      { id: 'mazeWall76', 
+      { id: 'mazeWall82', 
         width: `${widthModifier * 10}px`, 
         height: `${heightModifier * 10}px`, 
         top: `${heightModifier * 0}px`, 
@@ -629,7 +629,7 @@ export function createLevel2(widthModifier,heightModifier,offscreen) {
       columns.appendChild(wall);
     });
 
-    const wallBlack2 = createElement('div', 'wallBlack1', ['wallSolid']);
+    const wallBlack2 = createElement('div', 'wallBlack2', ['wallSolid']);
 
     const wallBlack2Config = [
       { id: 'wall1', 
@@ -709,6 +709,15 @@ export function createLevel2(widthModifier,heightModifier,offscreen) {
       lakeYellow.appendChild(lake);
     });
 
+    const gate2 = createElement('div', 'gate2', ['gate2'], {
+      background: 'green',
+      width: `${widthModifier * 10}px`,
+      height: `${heightModifier * 30}px`,
+      top: `${heightModifier * 220}px`,
+      left: `${widthModifier * 430}px`,
+    });
+    gate2.classList.add('wallSolid');
+
     // TODOconst shadows = createElement('div', 'shadow', ['shadow']);
 
     // const shadowsConfig = [
@@ -742,6 +751,7 @@ export function createLevel2(widthModifier,heightModifier,offscreen) {
   level2.appendChild(columns);
   level2.appendChild(wallBlack2);
   level2.appendChild(lakeYellow);
+  level2.appendChild(gate2);
   //TODO:level2.appendChild(shadows);
 
   playArea.appendChild(level2);
@@ -791,5 +801,8 @@ export function getLevel2Objects() {
   return childStack;
 }
 
-window.createLevel2 = createLevel2;
-window.createLevel2End = createLevel2End;
+export function openGateTwo() {
+  let gate2 = document.querySelector('#gate2');
+  gate2.style.transform = 'rotate(-180deg)';
+  gate2.style.transformOrigin = 'bottom left'; 
+}